requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
dataclasses-json==0.6.6
distro==1.9.0
firebase-admin==6.5.0
Flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
fuzzywuzzy==0.18.0
gcloud==0.18.3
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.5.4
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.64.0
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
jsonpatch==1.33
jsonpointer==2.4
jwcrypto==1.5.6
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.24
langchain-google-community==1.0.7
langchain-google-genai==1.0.5
langchain-text-splitters==0.2.0
langsmith==0.1.93
MarkupSafe==2.1.5
marshmallow==3.21.2
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauth2client==4.1.3
openai==1.58.1
orjson==3.10.3
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
Pyrebase4==4.8.0
python-dotenv==1.0.1
python-jwt==4.1.0
PyYAML==6.0.1
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.30
tenacity==8.3.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.0
uritemplate==4.1.1
urllib3==1.26.18
Werkzeug==3.0.3
yarl==1.9.4
youtube-search==2.1.2
